saco el archivo metodos de pago y lo pongo todo dentro de finalizar compra.
