+++ Frontend_API_TPO/src/pages/FinalizarCompra.jsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
-import { Container, Row, Col, Button } from 'react-bootstrap';
+import { Container, Row, Col, Button, Form, Image, Alert } from 'react-bootstrap';
 import { useNavigate } from 'react-router-dom';
-import MetodosDePago from '../components/MetodosDePago';
+import qr from '../assets/QRNAIKI.png';
 import '../styles/finalizarCompra.css';
 
 const FinalizarCompra = ({ formData }) => {
@@ -11,7 +11,6 @@ const FinalizarCompra = ({ formData }) => {
   const [discount] = useState(0); // Asumiendo que el descuento es 0 o se pasa de alguna manera
 
   const handleConfirmPurchase = () => {
-    // Lógica para realizar la compra (ej. enviar los datos al backend)
     alert('¡Compra realizada con éxito!');
     navigate('/');
   };
@@ -46,8 +45,12 @@ const FinalizarCompra = ({ formData }) => {
                 <option value="billeteraDigital">Billetera Digital</option>
                 <option value="transferenciaBancaria">Transferencia Bancaria</option>
               </select>
-              
-              {metodoPago && <MetodosDePago metodoPagoSeleccionado={metodoPago} />}
+
+              <div className="mt-4">
+                {metodoPago === 'tarjetaCredito' && <TarjetaCreditoForm onConfirm={handleConfirmPurchase} />}
+                {metodoPago === 'billeteraDigital' && <BilleteraDigitalForm onConfirm={handleConfirmPurchase} />}
+                {metodoPago === 'transferenciaBancaria' && <TransferenciaBancariaForm onConfirm={handleConfirmPurchase} />}
+              </div>
             </div>
           </Col>
           <Col md={4} className="bg-light p-3">
@@ -63,5 +66,155 @@ const FinalizarCompra = ({ formData }) => {
     </Container>
   );
 };
+
+// Formulario de Tarjeta de Crédito
+const TarjetaCreditoForm = ({ onConfirm }) => {
+  const [formData, setFormData] = useState({
+    numeroTarjeta: '',
+    fechaVencimiento: '',
+    codigoSeguridad: '',
+    nombreApellido: '',
+  });
+  const [errors, setErrors] = useState({});
+
+  const handleChange = (e) => {
+    const { name, value } = e.target;
+    setFormData({ ...formData, [name]: value });
+    if (value && errors[name]) {
+      setErrors((prevErrors) => {
+        const updatedErrors = { ...prevErrors };
+        delete updatedErrors[name];
+        return updatedErrors;
+      });
+    }
+  };
+
+  const validateFields = () => {
+    const newErrors = {};
+    Object.keys(formData).forEach((field) => {
+      if (!formData[field]) {
+        newErrors[field] = '*Obligatorio';
+      }
+    });
+
+    setErrors(newErrors);
+    if (Object.keys(newErrors).length === 0) {
+      onConfirm();
+    }
+  };
+
+  return (
+    <Form onSubmit={(e) => { e.preventDefault(); validateFields(); }}>
+      <Row>
+        <Col md={6}>
+          <Form.Group controlId="numeroTarjeta">
+            <Form.Label>Número de Tarjeta</Form.Label>
+            <Form.Control
+              type="text"
+              name="numeroTarjeta"
+              value={formData.numeroTarjeta}
+              onChange={handleChange}
+              placeholder="Ej. 1234 5678 9012 3456"
+              isInvalid={!!errors.numeroTarjeta}
+            />
+            <Form.Control.Feedback type="invalid">
+              {errors.numeroTarjeta}
+            </Form.Control.Feedback>
+          </Form.Group>
+        </Col>
+        <Col md={6}>
+          <Form.Group controlId="fechaVencimiento">
+            <Form.Label>Fecha de Vencimiento</Form.Label>
+            <Form.Control
+              type="text"
+              name="fechaVencimiento"
+              value={formData.fechaVencimiento}
+              onChange={handleChange}
+              placeholder="MM/AA"
+              isInvalid={!!errors.fechaVencimiento}
+            />
+            <Form.Control.Feedback type="invalid">
+              {errors.fechaVencimiento}
+            </Form.Control.Feedback>
+          </Form.Group>
+        </Col>
+      </Row>
+      <Row>
+        <Col md={6}>
+          <Form.Group controlId="codigoSeguridad">
+            <Form.Label>Código de Seguridad</Form.Label>
+            <Form.Control
+              type="text"
+              name="codigoSeguridad"
+              value={formData.codigoSeguridad}
+              onChange={handleChange}
+              placeholder="Ej. 123"
+              isInvalid={!!errors.codigoSeguridad}
+            />
+            <Form.Control.Feedback type="invalid">
+              {errors.codigoSeguridad}
+            </Form.Control.Feedback>
+          </Form.Group>
+        </Col>
+        <Col md={6}>
+          <Form.Group controlId="nombreApellido">
+            <Form.Label>Nombre y Apellido</Form.Label>
+            <Form.Control
+              type="text"
+              name="nombreApellido"
+              value={formData.nombreApellido}
+              onChange={handleChange}
+              placeholder="Ej. Juan Pérez"
+              isInvalid={!!errors.nombreApellido}
+            />
+            <Form.Control.Feedback type="invalid">
+              {errors.nombreApellido}
+            </Form.Control.Feedback>
+          </Form.Group>
+        </Col>
+      </Row>
+      <Button variant="success" type="submit" className="mt-3">
+        Confirmar Compra
+      </Button>
+    </Form>
+  );
+};
+
+// Formulario de Billetera Digital
+const BilleteraDigitalForm = ({ onConfirm }) => (
+  <div>
+    <h3>Último Paso: Realizar el Pago con tu Billetera Digital</h3>
+    <h5>Escanea el código QR y selecciona el método de pago de tu billetera</h5>
+    <Image src={qr} alt="Código QR" fluid />
+    <Form.Group className="mt-3">
+      <Form.Label>Adjunta el comprobante de pago</Form.Label>
+      <Form.Control type="file" />
+    </Form.Group>
+    <Button variant="success" onClick={onConfirm}>
+      Confirmar Compra
+    </Button>
+  </div>
+);
+
+// Formulario de Transferencia Bancaria
+const TransferenciaBancariaForm = ({ onConfirm }) => (
+  <div>
+    <h3>Último Paso: Realizar el Pago con Transferencia Bancaria</h3>
+    <Alert variant="secondary">
+      <p><strong>CBU:</strong> 2541002333058810</p>
+      <p><strong>Usuario:</strong> NaikiiZapas </p>
+      <p><strong>Empresa:</strong> Naikii </p>
+      <p><strong>Alias:</strong> naikii.ok </p>
+      <p><strong>Banco:</strong> Banco de la Nacion Argentina </p>
+    </Alert>
+    <Form.Group className="mt-3">
+      <Form.Label>Adjunta el comprobante de pago</Form.Label>
+      <Form.Control type="file" />
+    </Form.Group>
+    <Button variant="success" onClick={onConfirm}>
+      Confirmar Compra
+    </Button>
+  </div>
+);
 
 export default FinalizarCompra;
